feat(login, register redesign. Sync boombox)

+++ client/addons/boombox.lua
@@ -2,6 +2,7 @@ if not Config.Boombox.Enable then return end
 CreateThread(function()
     Info('Boombox is enabled')
 end)
+
 local boombox_ = joaat('prop_boombox_01')
 local carry_anim_dict = 'anim@heists@box_carry@'
 local carry_anim_name = 'idle'
@@ -21,6 +22,24 @@ local function nearbyBoombox()
     return false
 end
 
+---@param entity number
+---@return boolean
+local function takeEntityOwnership(entity)
+    if not DoesEntityExist(entity) then return false end
+    local player = PlayerId()
+    local netID = NetworkGetNetworkIdFromEntity(entity)
+    SetNetworkIdCanMigrate(netID, true)
+    SetNetworkIdExistsOnAllMachines(netID, true)
+    NetworkSetNetworkIdDynamic(netID, true)
+    SetNetworkIdSyncToPlayer(netID, player, true)
+    local time = GetGameTimer()
+    while NetworkGetEntityOwner(entity) ~= player and GetGameTimer() - time < 5000 do
+        NetworkRequestControlOfEntity(entity)
+        Wait(30)
+    end
+    return NetworkGetEntityOwner(entity) == player
+end
+
 local radioSettings = {
     silent = true
 }
@@ -29,6 +48,10 @@ local function openUi()
     local boombox = nearbyBoombox()
     if not boombox then return end
     local currentBoombox = boombox
+    if not NetworkGetEntityIsNetworked(currentBoombox) then return Warning('Boombox is not networked') end
+    local netId = NetworkGetNetworkIdFromEntity(currentBoombox)
+    Debug('netId', netId)
+    radioSettings.customId = netId
     OpenAudioPlayer(radioSettings, {
         onPlay = function(sound)
             if not DoesEntityExist(currentBoombox) then
@@ -37,6 +60,27 @@ local function openUi()
             end
             local volume = AudioVolume
             TriggerServerEvent('mx-audioplayer:attach', sound.soundId, NetworkGetNetworkIdFromEntity(currentBoombox), volume)
+        end,
+        onLogin = function(soundId, token)
+            Entity(currentBoombox).state:set('audioplayer_account', token, true)
+        end,
+        onLogout = function(soundId)
+            Entity(currentBoombox).state:set('audioplayer_account', nil, true)
+        end,
+        autoLogin = function(soundId)
+            local entity = Entity(currentBoombox)
+            if not entity.state.audioplayer_account then
+                return Debug('Auto login: No account found')
+            end
+            local account = entity.state.audioplayer_account
+            local success = Login({
+                token = account
+            })
+            if not success then
+                Surround:pushNotification('This user credentials has been modified. Please log in again.')
+                entity.state:set('audioplayer_account', nil, true)
+            end
+            Debug('Auto Login: success state', success)
         end
     })
 end
@@ -52,9 +96,9 @@ end
 local function create()
     local boombox = nearbyBoombox()
     if boombox then return end
-    local player = PlayerPed
-    local playerCoords = GetEntityCoords(player)
-    local heading = GetEntityHeading(player)
+    local ped = PlayerPed
+    local playerCoords = GetEntityCoords(ped)
+    local heading = GetEntityHeading(ped)
     RequestModel(boombox_)
     while not HasModelLoaded(boombox_) do
         RequestModel(boombox_)
@@ -66,16 +110,17 @@ local function create()
     SetEntityAsMissionEntity(object, true, true)
     SetModelAsNoLongerNeeded(boombox_)
     loadAnim(put_anim_dict)
-    TaskPlayAnim(player, put_anim_dict, put_anim_name, 8.0, -8.0, -1, 0, 0, false, false, false)
+    TaskPlayAnim(ped, put_anim_dict, put_anim_name, 8.0, -8.0, -1, 0, 0, false, false, false)
     Wait(200)
     PlaceObjectOnGroundProperly(object)
-    StopAnimTask(player, put_anim_dict, put_anim_name, 3.0)
+    StopAnimTask(ped, put_anim_dict, put_anim_name, 3.0)
 end
 
 RegisterNetEvent('mx-audioplayer:boombox:create', create)
 
 local function carryAnim()
     local ped = PlayerPed
+    if not takeEntityOwnership(ped) then return end
     loadAnim(carry_anim_dict)
     while carrying_boombox do
         if IsEntityPlayingAnim(PlayerPed, carry_anim_dict, carry_anim_name, 3) then goto continue end
@@ -91,6 +136,7 @@ local function drop()
     local playerCoords = GetEntityCoords(player)
     local boombox = nearbyBoombox()
     if not boombox then return end
+    if not takeEntityOwnership(boombox) then return end
     carrying_boombox = false
     DetachEntity(boombox, true, true)
     SetEntityCoords(boombox, playerCoords.x, playerCoords.y, playerCoords.z - 0.95, true, true, true, true)
@@ -105,6 +151,7 @@ local function pickup()
     local player = PlayerPed
     local boombox = nearbyBoombox()
     if not boombox then return end
+    if not takeEntityOwnership(boombox) then return end
     AttachEntityToEntity(boombox, player, GetPedBoneIndex(player, 24817), 0.0, 0.40, -0.0, -180.0, 90.0, 0.0, false, false, false, false, 2, true)
     carrying_boombox = true
     CreateThread(carryAnim)
@@ -123,6 +170,7 @@ end
 local function destroy()
     local boombox = nearbyBoombox()
     if not boombox then return end
+    if not takeEntityOwnership(boombox) then return end
     SetEntityAsMissionEntity(boombox, true, true)
     DeleteEntity(boombox)
     if DoesEntityExist(boombox) then return Debug('Boombox not deleted') end
